refactor: lnurl logic
